+++ examples/python/17_teleoperation_with_joint_mapping.py
@@ -1,7 +1,7 @@
 """
 Teleoperation Example
 
-Run this example on UPC to which the master arm and hands are connected
+Run this example on UPC to which the leader arm and hands are connected
 """
 
 import rby1_sdk as rby
@@ -31,7 +31,7 @@ class Pose:
 
 
 class Settings:
-    master_arm_loop_period = 1 / 100
+    leader_arm_loop_period = 1 / 100
 
     impedance_stiffness = 50
     impedance_damping_ratio = 1.0
@@ -269,7 +269,7 @@ def main(address, model, power, servo, control_mode):
         nonlocal robot_q
         robot_q = state.position
 
-    robot.start_state_update(robot_state_callback, 1 / Settings.master_arm_loop_period)
+    robot.start_state_update(robot_state_callback, 1 / Settings.leader_arm_loop_period)
 
     # ===== SETUP GRIPPER =====
     gripper = Gripper()
@@ -281,16 +281,16 @@ def main(address, model, power, servo, control_mode):
     gripper.homing()
     gripper.start()
 
-    # ===== SETUP MASTER ARM =====
-    rby.upc.initialize_device(rby.upc.MasterArmDeviceName)
-    master_arm_model = f"{os.path.dirname(os.path.realpath(__file__))}/../../models/master_arm/model.urdf"
-    master_arm = rby.upc.MasterArm(rby.upc.MasterArmDeviceName)
-    master_arm.set_model_path(master_arm_model)
-    master_arm.set_control_period(Settings.master_arm_loop_period)
-    active_ids = master_arm.initialize(verbose=False)
-    if len(active_ids) != rby.upc.MasterArm.DeviceCount:
+    # ===== SETUP LEADER ARM =====
+    rby.upc.initialize_device(rby.upc.LeaderArmDeviceName)
+    leader_arm_model = f"{os.path.dirname(os.path.realpath(__file__))}/../../models/leader_arm/model.urdf"
+    leader_arm = rby.upc.LeaderArm(rby.upc.LeaderArmDeviceName)
+    leader_arm.set_model_path(leader_arm_model)
+    leader_arm.set_control_period(Settings.leader_arm_loop_period)
+    active_ids = leader_arm.initialize(verbose=False)
+    if len(active_ids) != rby.upc.LeaderArm.DeviceCount:
         logging.error(
-            f"Mismatch in the number of devices detected for RBY Master Arm (active devices: {active_ids})"
+            f"Mismatch in the number of devices detected for RBY Leader Arm (active devices: {active_ids})"
         )
         exit(1)
 
@@ -319,7 +319,7 @@ def main(address, model, power, servo, control_mode):
 
     log_count = 0
 
-    def master_arm_control_loop(state: rby.upc.MasterArm.State):
+    def leader_arm_control_loop(state: rby.upc.LeaderArm.State):
         nonlocal position_mode, right_q, left_q, right_minimum_time, left_minimum_time, log_count
 
         if right_q is None:
@@ -327,10 +327,10 @@ def main(address, model, power, servo, control_mode):
         if left_q is None:
             left_q = state.q_joint[7:14]
 
-        ma_input = rby.upc.MasterArm.ControlInput()
+        ma_input = rby.upc.LeaderArm.ControlInput()
 
         log_count += 1
-        if log_count % round(1 / Settings.master_arm_loop_period) == 0:
+        if log_count % round(1 / Settings.leader_arm_loop_period) == 0:
             print(f"--- {datetime.datetime.now().time()} ---")
             print(f"Button: {state.button_right.button}, {state.button_left.button}")
             print(f"Trigger: {state.button_right.trigger}, {state.button_left.trigger}")
@@ -340,7 +340,7 @@ def main(address, model, power, servo, control_mode):
                 [state.button_right.trigger / 1000, state.button_left.trigger / 1000]
             )
         )
-        # ===== CALCULATE MASTER ARM COMMAND =====
+        # ===== CALCULATE LEADER ARM COMMAND =====
         torque = (
             state.gravity_term
             + ma_q_limit_barrier
@@ -391,9 +391,9 @@ def main(address, model, power, servo, control_mode):
         # ===== BUILD ROBOT COMMAND =====
         rc = rby.BodyComponentBasedCommandBuilder()
         if state.button_right.button and not is_collision:
-            right_minimum_time -= Settings.master_arm_loop_period
+            right_minimum_time -= Settings.leader_arm_loop_period
             right_minimum_time = max(
-                right_minimum_time, Settings.master_arm_loop_period * 1.01
+                right_minimum_time, Settings.leader_arm_loop_period * 1.01
             )
             right_arm_builder = (
                 rby.JointPositionCommandBuilder()
@@ -430,9 +430,9 @@ def main(address, model, power, servo, control_mode):
             right_minimum_time = 0.8
 
         if state.button_left.button and not is_collision:
-            left_minimum_time -= Settings.master_arm_loop_period
+            left_minimum_time -= Settings.leader_arm_loop_period
             left_minimum_time = max(
-                left_minimum_time, Settings.master_arm_loop_period * 1.01
+                left_minimum_time, Settings.leader_arm_loop_period * 1.01
             )
             left_arm_builder = (
                 rby.JointPositionCommandBuilder()
@@ -475,12 +475,12 @@ def main(address, model, power, servo, control_mode):
 
         return ma_input
 
-    master_arm.start_control(master_arm_control_loop)
+    leader_arm.start_control(leader_arm_control_loop)
 
     # ===== SETUP SIGNAL =====
     def handler(signum, frame):
         robot.stop_state_update()
-        master_arm.stop_control()
+        leader_arm.stop_control()
         robot.cancel_control()
         time.sleep(0.5)